requirements.txt
pynput==1.7.6
Pillow==10.0.0
pystray==0.19.4
pyinstaller==6.3.0 